test: move GetFreePort into test/framework/net
GetFreePort is a network helper, not a general test util. Moving it into the
test/framework tree gives it an obvious home and keeps test/utils from growing
into a grab bag. The function body is unchanged; the only edit is the package
clause plus a license header and package comment on the new file.

Part of #1188
