+++ test/framework/net/net.go
@@ -1,9 +1,21 @@
-// Package utils contains utilities for testing
-//
-//revive:disable:var-naming
-package utils
+/*
+Copyright 2026 The llm-d Authors.
 
-//revive:enable:var-naming
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+// Package net provides network helpers for tests.
+package net
 
 import (
 	"errors"
